+++ NEWS.md
@@ -4,8 +4,9 @@
 ## v0.5.0
 
 - Now that `@__DIR__` is provided, do not `cd()` into the test directory.
-- Allow running tests in the context of a specific module.
+- Added an option to run tests in the context of a specific module.
 - Added an option to suppress any output except for error reports.
+- Generate valid names for test modules.
 
 
 ## v0.4.1